requirements.txt
allure-pytest==2.9.41
allure-python-commons==2.9.41
pytest==6.2.4
pytest-rerunfailures==10.0
selene==1.0.2
selenium==3.141.0
webdriver-manager==3.4.2
